+++ tests/testthat/test-roxytest-testexamples-sig_fit.R
@@ -2,7 +2,7 @@
 
 context("File R/sig_fit.R: @testexamples")
 
-test_that("Function sig_fit() @ L95", {
+test_that("Function sig_fit() @ L97", {
   
   W <- matrix(c(1, 2, 3, 4, 5, 6), ncol = 2)
   colnames(W) <- c("sig1", "sig2")